pyproject.toml
[project]
name = "bobafinder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "googlemaps>=4.10.0",
    "langchain>=1.2.3",
    "langchain-fireworks>=1.1.0",
    "langgraph-checkpoint-mongodb>=0.3.0",
    "langgraph-swarm>=0.1.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pymongo[srv]>=4.15.5",
    "python-dotenv>=1.2.1",
    "yelpapi>=2.0.0",
]
